Split out graph types, rename vertical and horizontal graphs and add character parameter.
